% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageBumpyDataFrameMatrix.R
\name{stageObject,BumpyDataFrameMatrix-method}
\alias{stageObject,BumpyDataFrameMatrix-method}
\title{Stage BumpyDataFrameMatrix objects}
\usage{
\S4method{stageObject}{BumpyDataFrameMatrix}(x, dir, path, child = FALSE)
}
\arguments{
\item{x}{A BumpyDataFrameMatrix object.}

\item{dir}{String containing the path to the staging directory.}

\item{path}{String containing a prefix of the relative path inside \code{dir} where \code{x} is to be saved.
The actual path used to save \code{x} may include additional components, see Details.}

\item{child}{Logical scalar indicating whether \code{x} is a child of a larger object.}
}
\value{
A named list of metadata, to be used by \code{\link{.writeMetadata}}.
}
\description{
Staging method for \linkS4class{BumpyDataFrameMatrix} objects.
See \code{?\link{stageObject}} for more details.
}
\seealso{
\code{\link{useBumpyHDF5}}, to control whether the concatenated data frame is saved as a CSV or HDF5 file.
}
\author{
Aaron Lun
}
