% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadBumpyDataFrameMatrix.R
\name{loadBumpyDataFrameMatrix}
\alias{loadBumpyDataFrameMatrix}
\title{Load a BumpyDataFrameMatrix object}
\usage{
loadBumpyDataFrameMatrix(assay.info, project)
}
\arguments{
\item{assay.info}{Named list of metadata for the BumpyDataFrameMatrix, generated by the corresponding \code{\link{stageObject}} method.}

\item{project}{Any argument accepted by the acquisition functions, see \code{?\link[alabaster.base]{acquireFile}}.
By default, this should be a string containing the path to a staging directory.}
}
\value{
A \linkS4class{BumpyDataFrameMatrix} object.
}
\description{
Load a \linkS4class{BumpyDataFrameMatrix} object based on its metadata.
}
\author{
Aaron Lun
}
